# openapi.yaml
openapi: 3.0.0
info:
  title: Hack Backend API
  description: |
    API for civic issue reporting and management system.
    - Users can report issues, view their issues, and repost issues
    - Admins can manage issues, update status, and add comments
    - Authentication via JWT tokens (Bearer token)
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
...
  /user/issues/{issueId}/comments:
    post:
      tags:
        - Comments
      summary: Create a comment on an issue (User)
      description: Add a comment to any issue (users can comment on any issue, not just their own)
      operationId: createUserComment
      security:
        - bearerAuth: []
      parameters:
        - name: issueId
          in: path
          required: true
          description: Issue ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: "This is a serious issue that needs attention!"
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  comment:
                    $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '404':
          description: Issue or user not found
        '500':
          description: Server error

    get:
      tags:
        - Comments
      summary: Get comments for an issue (User)
      description: Retrieve all comments for a specific issue (users can view comments on any issue)
      operationId: getUserComments
      security:
        - bearerAuth: []
      parameters:
        - name: issueId
          in: path
          required: true
          description: Issue ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Comments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  comments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
                  count:
                    type: integer
        '401':
          description: Unauthorized
        '404':
          description: Issue not found
        '500':
          description: Server error
  /admin/issues/{issueId}/comments:
    post:
      tags:
        - Comments
      summary: Create a comment on an issue (Admin)
      description: Add a comment to an issue
      operationId: createComment
      security:
        - bearerAuth: []
      parameters:
        - name: issueId
          in: path
          required: true
          description: Issue ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: "We are investigating this issue and will update you soon."
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  comment:
                    $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
        '404':
          description: Issue or admin not found
        '500':
          description: Server error

    get:
      tags:
        - Comments
      summary: Get comments for an issue (Admin)
      description: Retrieve all comments for a specific issue
      operationId: getAdminComments
      security:
        - bearerAuth: []
      parameters:
        - name: issueId
          in: path
          required: true
          description: Issue ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Comments retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  comments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Comment'
                  count:
                    type: integer
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required
        '404':
          description: Issue not found
        '500':
          description: Server error
  /admin/comments/{commentId}:
    put:
      tags:
        - Comments
      summary: Update a comment (Admin)
      description: Update a comment (can only update own comments)
      operationId: updateComment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          description: Comment ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  example: "Updated comment content"
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  comment:
                    $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Can only update own comments
        '404':
          description: Comment not found
        '500':
          description: Server error

    delete:
      tags:
        - Comments
      summary: Delete a comment (Admin)
      description: Delete a comment (can only delete own comments)
      operationId: deleteComment
      security:
        - bearerAuth: []
      parameters:
        - name: commentId
          in: path
          required: true
          description: Comment ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Can only delete own comments
        '404':
          description: Comment not found
        '500':
          description: Server error

  # Repost Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoints
  schemas:
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        content:
          type: string
        issueId:
          type: string
          format: uuid
        issue:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            title:
              type: string
        adminId:
          type: string
          format: uuid
          nullable: true
          description: Present if comment is from an admin
        admin:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            email:
              type: string
              format: email
        userId:
          type: string
          format: uuid
          nullable: true
          description: Present if comment is from a user
        user:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            number:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...